requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-pptx==0.6.23
Werkzeug==3.0.1
Pillow==10.1.0
python-dotenv==1.0.0
bcrypt==4.1.2